# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PolyGo API",
    "description": "High-performance Polymarket API proxy with caching and WebSocket support",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@polygo.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "POLY-API-KEY"
    }
  },
  "paths": {
...
    "/api/v1/price/{token_id}": {
      "get": {
        "tags": ["Prices"],
        "summary": "Get current price",
        "produces": ["application/json"],
        "parameters": [
          {"name": "token_id", "in": "path", "required": true, "type": "string"},
          {"name": "side", "in": "query", "type": "string", "default": "BUY"}
        ],
        "responses": {
          "200": {"description": "Current price"}
        }
      }
    },
    "/api/v1/book/{token_id}": {
      "get": {
        "tags": ["Prices"],
        "summary": "Get order book",
        "produces": ["application/json"],
        "parameters": [
          {"name": "token_id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {
          "200": {"description": "Order book"}
        }
      }
    },
    "/api/v1/spread/{token_id}": {
      "get": {
        "tags": ["Prices"],
        "summary": "Get spread",
        "produces": ["application/json"],
        "parameters": [
          {"name": "token_id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {
          "200": {"description": "Spread info"}
        }
      }
    },
    "/api/v1/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get user orders",
        "security": [{"ApiKeyAuth": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {"description": "List of orders"},
          "401": {"description": "Unauthorized"}
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create order",
        "security": [{"ApiKeyAuth": []}],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "order",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "tokenID": {"type": "string"},
                "side": {"type": "string", "enum": ["BUY", "SELL"]},
                "price": {"type": "string"},
                "size": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "200": {"description": "Order created"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/api/v1/orders/{id}": {
      "delete": {
        "tags": ["Orders"],
        "summary": "Cancel order",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {
          "200": {"description": "Order cancelled"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
...
  },
...
}